move print_separated utilities into common crate
